add heartbeat logic
Following the pattern from the ruby implementation, spawn two babysitting
threads, one to see for outstanding sends, and one for outstanding recvs

Sadly this means we need to mutex the global state.  Game-looping
this to the end of the inner loop of main-loop may be an acceptable 'gc
pause' instead of this approach.